a better timer trait, a resampler and no clue about dependant probabilities.

+++ mimir/models/Sample.h
@@ -21,7 +21,7 @@ public:
 private:
     ValueIndex _classifier;
     ValueIndex _valueIndex;
-    unsigned _count;
+    unsigned long _count;
 };
 
 } // namespace models
